requirements.txt
pandas
pyserial
numpy